requirements/prod.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --allow-unsafe requirements/prod.in
#
appdirs==1.4.4
    # via
    #   cumulusci
    #   fs
authlib==0.15.4
    # via
    #   cumulusci
    #   simple-salesforce
certifi==2021.5.30
    # via
    #   cumulusci
    #   requests
cffi==1.14.5
    # via
    #   cryptography
    #   cumulusci
chardet==4.0.0
    # via
    #   cumulusci
    #   requests
click==7.1.2
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   snowfakery
coloredlogs==15.0
    # via cumulusci
contextvars==2.4
    # via -r requirements/prod.in
cryptography==3.4.7
    # via
    #   authlib
    #   cumulusci
    #   jwcrypto
cumulusci==3.37.0
    # via -r requirements/prod.in
deprecated==1.2.12
    # via
    #   cumulusci
    #   jwcrypto
docutils==0.16
    # via cumulusci
faker==8.6.0
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   snowfakery
fs==2.4.13
    # via cumulusci
github3.py==2.0.0
    # via cumulusci
greenlet==1.1.0
    # via
    #   cumulusci
    #   sqlalchemy
gvgen==1.0
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   snowfakery
humanfriendly==9.1
    # via
    #   coloredlogs
    #   cumulusci
idna==2.10
    # via
    #   cumulusci
    #   requests
immutables==0.15
    # via contextvars
importlib-metadata==4.5.0
    # via
    #   cumulusci
    #   keyring
    #   sqlalchemy
jinja2==2.11.3
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   snowfakery
jwcrypto==0.9
    # via
    #   cumulusci
    #   github3.py
keyring==23.0.1
    # via cumulusci
lxml==4.6.3
    # via cumulusci
markupsafe==1.1.1
    # via
    #   cumulusci
    #   jinja2
pycparser==2.20
    # via
    #   cffi
    #   cumulusci
pydantic==1.8.2
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   snowfakery
pyjwt==2.1.0
    # via cumulusci
python-dateutil==2.8.1
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   faker
    #   github3.py
    #   snowfakery
pytz==2021.1
    # via
    #   cumulusci
    #   fs
pyyaml==5.4.1
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   snowfakery
requests-futures==1.0.0
    # via cumulusci
requests==2.25.1
    # via
    #   cumulusci
    #   github3.py
    #   requests-futures
    #   robotframework-requests
    #   salesforce-bulk
    #   simple-salesforce
robotframework-lint==1.1
    # via cumulusci
robotframework-pabot==2.0.0
    # via cumulusci
robotframework-pythonlibcore==2.2.1
    # via
    #   cumulusci
    #   robotframework-seleniumlibrary
robotframework-requests==0.9.1
    # via cumulusci
robotframework-seleniumlibrary==5.1.3
    # via cumulusci
robotframework-stacktrace==0.4.0
    # via
    #   cumulusci
    #   robotframework-pabot
robotframework==4.0.3
    # via
    #   cumulusci
    #   robotframework-lint
    #   robotframework-pabot
    #   robotframework-requests
    #   robotframework-seleniumlibrary
    #   robotframework-stacktrace
rst2ansi==0.1.5
    # via cumulusci
salesforce-bulk==2.2.0
    # via cumulusci
sarge==0.1.6
    # via cumulusci
selenium==3.141.0
    # via
    #   cumulusci
    #   robotframework-seleniumlibrary
simple-salesforce==1.11.1
    # via
    #   cumulusci
    #   salesforce-bulk
six==1.16.0
    # via
    #   cumulusci
    #   fs
    #   jwcrypto
    #   python-dateutil
    #   salesforce-bulk
snowfakery==1.12
    # via cumulusci
sqlalchemy==1.4.17
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   snowfakery
terminaltables==3.1.0
    # via cumulusci
text-unidecode==1.3
    # via
    #   cumulusci
    #   faker
typing-extensions==3.10.0.0
    # via
    #   -r requirements/prod.in
    #   cumulusci
    #   importlib-metadata
    #   pydantic
    #   snowfakery
unicodecsv==0.14.1
    # via
    #   cumulusci
    #   salesforce-bulk
uritemplate==3.0.1
    # via
    #   cumulusci
    #   github3.py
urllib3==1.26.5
    # via
    #   cumulusci
    #   requests
    #   selenium
wrapt==1.12.1
    # via
    #   cumulusci
    #   deprecated
xmltodict==0.12.0
    # via cumulusci
zipp==3.4.1
    # via
    #   cumulusci
    #   importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
setuptools==57.0.0
    # via
    #   cumulusci
    #   fs
